+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Changed
+* `cargoNextest` now supports setting `withLlvmCov` which will automatically run
+  `cargo llvm-cov nextest`. Note that `withLlvmCov = true;` is (currently) only
+  supported when `partitions = 1;`
+
 ## [0.16.3] - 2024-03-19
 
 ### Changed